Move the check whether a chapter is deleted before purging to the service

+++ src/Service/ComicsService.php
@@ -18,6 +18,9 @@ class ComicsService
 
     public function purge(Chapter $chapter)
     {
+        if (!$chapter->getIsDeleted()) {
+            throw new \Exception('Cannot purge a chapter that is not deleted!');
+        }
         $this->entityManager->remove($chapter);
         $this->entityManager->flush();
         $filesystem = new Filesystem();
